fix for teach page
