+++ src/components/teach/manageClassrooms/components/CreateClassDialog.scss
@@ -256,6 +256,10 @@
         }
       }
 
+      #classroom-code-vr3 {
+        text-transform: uppercase;
+      }
+
       .copy-btn {
         transition: all 300ms ease-in-out;
         cursor: pointer;
